requirements.txt
amqp==5.3.1
anyio==4.11.0
asgiref==3.9.1
billiard==4.2.1
cachetools==5.5.2
celery==5.5.3
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cron_descriptor==2.0.6
distro==1.9.0
Django==5.2.6
django-celery-beat==2.8.1
django-timezone-field==7.1
django_celery_results==2.6.0
djangorestframework==3.16.1
google-api-core==2.25.1
google-api-python-client==2.181.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
googleapis-common-protos==1.70.0
groq==0.32.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.0
httpx==0.28.1
idna==3.10
kombu==5.5.4
oauthlib==3.3.1
packaging==25.0
prompt_toolkit==3.0.52
proto-plus==1.26.1
protobuf==6.32.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==1.10.24
pyparsing==3.2.4
python-crontab==3.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
redis==6.4.0
requests==2.32.5
requests-oauthlib==2.0.0
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.13
